backend/pyproject.toml
[project]
name = "joysafeter"
version = "0.1.0"
description = "JoySafeter Backend - Intelligent Agent Platform Backend Service"
requires-python = ">=3.12,<3.14"
license = {text = "Apache-2.0"}
authors = [
    {name = "JoySafeter Team", email = "team@joysafeter.com"}
]
keywords = ["agent", "langchain", "langgraph", "ai", "platform"]

dependencies = [
    # Web Framework
    "fastapi[standard]>=0.122.0",
    "uvicorn[standard]>=0.38.0",
    "gunicorn>=23.0.0",
    "python-multipart>=0.0.12",
    "sse-starlette>=2.0.0",
    "hypercorn>=0.17.3",
    # Database
    "sqlalchemy>=2.0.44",
    "asyncpg>=0.30.0",
    "psycopg[binary,pool]>=3.1.0", # Psycopg 3 with binary for langgraph-checkpoint-postgres
    "psycopg2-binary>=2.9.11", # Psycopg 2 for legacy scripts
    "alembic>=1.14.0",
    "greenlet>=3.0.0", # SQLAlchemy async support
    # LangChain & LangGraph
    "langchain-openai>=1.0.0",
    "langgraph-checkpoint>=2.0.0",
    "langgraph-checkpoint-postgres>=3.0.2",
    "langchain-mcp-adapters>=0.1.13",
    "deepagents>=0.4.0",
    "pydantic-ai-backend[docker]>=0.1.14",
    # MCP
    "mcp>=1.20.0",
    "fastmcp>=2.14.1",
    # Observability
    "langfuse>=3.11.1",
    "loguru>=0.7.3",
    # Auth & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=5.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
    # HTTP & Networking
    "aiohttp>=3.13.2",
    "httpx>=0.27.0",
    "requests>=2.32.5",
    "websockets>=13.0",
    # Utils
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
    "aiofiles>=24.0.0",
    "orjson>=3.10.0",
    "rich>=13.7.0",
    "python-frontmatter>=1.1.0",
    "markdownify>=0.11.6",
    # Email
    "aiosmtplib>=3.0.0",
    # Caching
    "redis>=5.0.0",
    # Web scraping & automation
    "beautifulsoup4>=4.14.2",
    "selenium>=4.38.0",
    "mitmproxy>=11.1.3",
    # Security analysis
    "angr>=9.2.182",
    # System & Docker
    "psutil>=7.1.3",
    "docker>=7.1.0",
    # Search
    "tavily-python>=0.3.0",
    "langchain-google-genai>=4.2.0",
    "langchain-anthropic>=1.3.3",
    "semver>=3.0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "httpx>=0.27.0",
    "types-PyYAML>=6.0.0",
    "types-redis>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "types-PyYAML>=6.0.0",
    "types-redis>=4.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"
src = [".", "src"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
# Do not treat unannotated function body notes as errors; avoids CI failure on notes
disable_error_code = ["annotation-unchecked"]

[tool.uv]
index-url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
